// src/app/models/customer.ts
export interface Customer {
  id?: number;
  customerCode: string;
  customerName: string;
  contactPerson?: string;
  mobileNo?: string;
  email?: string;
  address1?: string;
  address2?: string;
  city?: string;
  state?: string;
  country?: string;
  zipCode?: string;
  gstNo?: string;
  isActive: boolean;
  isDeleted?: boolean;
  createdBy?: number;
  createdDate?: Date;
  updatedBy?: number;
  updatedDate?: Date;
}